add logic to cafe, food and drink controllers, add some field to DB edit models add controller to edit password and name cafe add tokinazer fix some bugs

+++ controllers/cafe/deleteCafe.js
@@ -1,4 +1,7 @@
-let dataBase = require('../../dataBase').getInstance();
+const dataBase = require('../../dataBase').getInstance();
+const tokenVerifikator = require('../../helpers/tokenVerificator');
+const secret = require('../../config/secret');
+const userRoles = require('../../config/userRoles');
 
 module.exports = async (req, res) => {
     try {
@@ -8,15 +11,22 @@ module.exports = async (req, res) => {
 
         if (!name) throw new Error('No name');
 
-        await Cafe.destroy({
-            where: {
-                name
-            }
-        });
+        const token = req.get("Authorization");
+        if (!token) throw new Error('No token');
 
+        const {admin} = userRoles;
+
+        const {name: nameFromToken} = tokenVerifikator(token, secret);
+        if (nameFromToken !== admin) throw  new Error('You are not admin');
+
+            await Cafe.destroy({
+                where: {
+                    name
+                }
+            });
         res.json({
             success: true,
-            message: 'Cafe successfully deleted'
+            message: `Cafe ${name} successfully deleted`
         });
     } catch (e) {
         console.log(e);
